requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
Pillow>=8.3.0
requests>=2.25.0
moxing>=2.0.0
